+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/action/BrowserAction.kt
@@ -33,6 +33,7 @@ import mozilla.components.browser.state.state.recover.RecoverableTab
 import mozilla.components.browser.state.state.recover.TabState
 import mozilla.components.concept.engine.Engine
 import mozilla.components.concept.engine.EngineSession
+import mozilla.components.concept.engine.EngineSession.CookieBannerHandlingStatus
 import mozilla.components.concept.engine.EngineSessionState
 import mozilla.components.concept.engine.HitResult
 import mozilla.components.concept.engine.content.blocking.Tracker
@@ -824,6 +825,18 @@ sealed class TrackingProtectionAction : BrowserAction() {
      * Clears the [TrackingProtectionState.blockedTrackers] and [TrackingProtectionState.blockedTrackers] lists.
      */
     data class ClearTrackersAction(val tabId: String) : TrackingProtectionAction()
+}
+
+/**
+ * [BrowserAction] implementations related to updating the [SessionState.cookieBanner] of a single [SessionState] inside
+ * [BrowserState].
+ */
+sealed class CookieBannerAction : BrowserAction() {
+    /**
+     * Updates the [SessionState.cookieBanner] state or a a single [SessionState].
+     */
+    data class UpdateStatusAction(val tabId: String, val status: CookieBannerHandlingStatus) :
+        CookieBannerAction()
 }
 
 /**