+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/CustomTabSessionState.kt
@@ -5,6 +5,7 @@
 package mozilla.components.browser.state.state
 
 import mozilla.components.concept.engine.EngineSession
+import mozilla.components.concept.engine.EngineSession.CookieBannerHandlingStatus
 import mozilla.components.concept.engine.manifest.WebAppManifest
 import java.util.UUID
 
@@ -32,6 +33,7 @@ data class CustomTabSessionState(
     override val contextId: String? = null,
     override val source: SessionState.Source = SessionState.Source.Internal.CustomTab,
     override val restored: Boolean = false,
+    override val cookieBanner: CookieBannerHandlingStatus = CookieBannerHandlingStatus.NO_DETECTED,
 ) : SessionState {
 
     override fun createCopy(
@@ -42,6 +44,7 @@ data class CustomTabSessionState(
         extensionState: Map<String, WebExtensionState>,
         mediaSessionState: MediaSessionState?,
         contextId: String?,
+        cookieBanner: CookieBannerHandlingStatus,
     ) = copy(
         id = id,
         content = content,